legger til OmOppholdetDittModal
